requirements.txt
matplotlib>=3.7.0
sentence-transformers==2.2.2
detoxify>=0.5.0
nltk>=3.8.1
spacy>=3.8.0
pillow>=10.0.0
pandas>=2.0.0
gradio==5.45.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers>=4.36.0
huggingface-hub==0.33.5
